pyproject.toml
[project]
name = "gestor-educativo"
version = "0.1.0"
description = "Backend para gestión de notas, diplomas y certificados académicos"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Gestor Educativo Team"}]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",
    # Document generation
    "docxtpl>=0.18.0",
    "python-docx>=1.1.0",
    "qrcode[pil]>=7.4",
    # OCR & Image processing
    "pytesseract>=0.3.10",
    "pdf2image>=1.17.0",
    "Pillow>=10.3.0",
    # LLM
    "openai>=1.30.0",
    # Email (Gmail API)
    "google-api-python-client>=2.130.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    # Logging
    "structlog>=24.1.0",
    # Settings
    "pydantic-settings>=2.2.0",
    # HTTP client
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "pip-audit>=2.7.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "S", "B", "A", "C4", "UP"]
ignore = ["S101"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
